chore: move Popup to popup package
