+++ app/src/main/java/fr/iut_rodez/pathpilot_android_client/home/itinerary/InfoItinerary.java
@@ -18,7 +18,7 @@ import fr.iut_rodez.pathpilot_android_client.home.clients.ClientArrayAdapter;
 import fr.iut_rodez.pathpilot_android_client.home.routes.IRouteService;
 import fr.iut_rodez.pathpilot_android_client.home.routes.Route;
 import fr.iut_rodez.pathpilot_android_client.login.JWTToken;
-import fr.iut_rodez.pathpilot_android_client.util.Popup;
+import fr.iut_rodez.pathpilot_android_client.util.popup.Popup;
 import fr.iut_rodez.pathpilot_android_client.util.popup.DialogButton;
 
 public class InfoItinerary extends AppCompatActivity {
