+++ app/src/main/java/fr/iut_rodez/pathpilot_android_client/login/Login.java
@@ -13,7 +13,7 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import fr.iut_rodez.pathpilot_android_client.R;
 import fr.iut_rodez.pathpilot_android_client.signup.SignUp;
-import fr.iut_rodez.pathpilot_android_client.util.Popup;
+import fr.iut_rodez.pathpilot_android_client.util.popup.Popup;
 import fr.iut_rodez.pathpilot_android_client.util.ValidateForm;
 import fr.iut_rodez.pathpilot_android_client.util.network.NetworkUtils;
