+++ app/src/main/java/fr/iut_rodez/pathpilot_android_client/signup/SignUp.java
@@ -30,7 +30,7 @@ import fr.iut_rodez.pathpilot_android_client.R;
 import fr.iut_rodez.pathpilot_android_client.login.Login;
 import fr.iut_rodez.pathpilot_android_client.map.MapSelection;
 import fr.iut_rodez.pathpilot_android_client.signup.SignUpService.SignUpInput;
-import fr.iut_rodez.pathpilot_android_client.util.Popup;
+import fr.iut_rodez.pathpilot_android_client.util.popup.Popup;
 
 /**
  * Handle the sign up Activity
